Final icon & link metadata separation
