+++ src/components/Header.tsx
@@ -4,7 +4,50 @@ import "./Header.scss"
 
 import { CreateIcon, CreateLink } from "./Link"
 import { useSiteMetadata } from "../hooks/use-site-metadata"
-import socialLinks from "../meta/links/social"
+
+const email = {
+  from: "Email",
+  title: "Enviame un Email",
+  href: "mailto:sudo@angelxehg.com",
+}
+
+const resume = {
+  from: "File",
+  title: "Descarga mi CV",
+  href: "https://drive.google.com/drive/folders/18DEv-IdP2veHzUINHafEk4Gz9i2zakzG?usp=sharing",
+}
+
+const devTo = {
+  from: "Dev.to",
+  title: "Visita mi blog en dev.to",
+  href: "https://dev.to/angelxehg",
+}
+
+const linkedIn = {
+  from: "LinkedIn",
+  title: "LinkedIn (/in/angelxehg)",
+  href: "https://www.linkedin.com/in/angelxehg",
+}
+
+const gitHub = {
+  from: "GitHub",
+  title: "GitHub (@angelxehg)",
+  href: "https://github.com/angelxehg",
+}
+
+const twitter = {
+  from: "Twitter",
+  title: "Twitter (@angelxehg)",
+  href: "https://www.twitter.com/angelxehg",
+}
+
+const instagram = {
+  from: "Instagram",
+  title: "Instagram (@angelxehg)",
+  href: "https://www.instagram.com/angelxehg",
+}
+
+const socialLinks = [email, resume, devTo, linkedIn, gitHub, twitter, instagram]
 
 const SocialLinks = () => (
   <nav>
