+++ src/components/Link.tsx
@@ -1,7 +1,8 @@
 import React from "react"
 
 import Icon, { IconProps } from "./Icon"
-import allLinks, { getLinkMeta, LinkMeta } from "../meta/links"
+import allLinks, { getLinkMeta } from "../meta/links"
+import { LinkMeta } from "../meta/links/types"
 
 const ClickableIcon = (props: {
   className?: string
